requirements.txt
streamlit
PyMuPDF
sumy
duckduckgo-search
python-pptx
nltk
